requirements.txt
fastapi
uvicorn
requests
pymupdf
tqdm
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
chromadb